Xu merge cpu backend code (#2747)
* update submodules for cpu device.
* update csrc/cpu code.
* update tests/cpu UTs.
* turn on cpu build.
* update cpu jit c++ code.
* pass down compiler cfg to cmake extenral project.

---------

+++ csrc/jit/cpu/tensorexpr/external_call/conv_add_relu.h
@@ -76,7 +76,7 @@ struct LoweringFuncTrait<ConvFusedOp::kConvAddReLU>
 
   static ideep::attr_t get_attr(int64_t* extra_args) {
     constexpr int alpha_idx = 0;
-    const float alpha = static_cast<float>(((int64_t*)extra_args)[alpha_idx]);
+    const float alpha = static_cast<float>(((double*)extra_args)[alpha_idx]);
     return ideep::attr_t::residual(alpha);
   }
